inverter aus UI installieren

+++ functions/storm.house.bash
@@ -303,6 +303,8 @@ setup_wb_config() {
     if [[ -z "${1:-$wallboxtype}" ]]; then echo "SKIPPED (no wallbox defined)"; return 1; fi
   fi
 
+   if ! cond_redirect install -m 755 "${sdIncludesDir}/setup_inverter" /usr/local/sbin; then echo "FAILED (install setup_inverter)"; return 1; fi
+
   for component in things items rules; do
     rm -f "${OPENHAB_CONF:-/etc/openhab}/${component}/wb.${component}"
     srcfile="${OPENHAB_CONF:-/etc/openhab}/${component}/STORE/${1:-${wallboxtype}}.${component}"
